chore(data-service): migrate data-service to typescript COMPASS-5009
