+++ packages/data-service/package.json
@@ -39,9 +39,10 @@
     "cover": "nyc npm run test",
     "depcheck": "depcheck",
     "lint": "eslint . && prettier --check .",
-    "prepare": "npm run compile",
+    "bootstrap": "npm run compile",
+    "prepublishOnly": "npm run compile",
     "pretest": "npm run compile && mongodb-runner start --port=27018",
-    "test": "npm run lint && nyc mocha --colors --timeout 15000 -r ts-node/register {src,test}/**/*.spec.*",
+    "test": "npm run lint && nyc mocha --colors --timeout 15000 -r ts-node/register src/**/*.spec.*",
     "test-ci": "npm run test",
     "test-check-ci": "npm run check && npm test",
     "posttest": "mongodb-runner stop --port=27018",
@@ -63,13 +64,14 @@
     "mongodb-index-model": "^3.4.0",
     "mongodb-js-errors": "^0.5.0",
     "mongodb-ns": "^2.2.0",
-    "mongodb-security": "^1.3.0",
-    "mongodb-url": "^3.0.3"
+    "mongodb-security": "^1.3.0"
   },
   "devDependencies": {
     "@mongodb-js/eslint-config-compass": "*",
     "@mongodb-js/prettier-config-compass": "*",
     "@mongodb-js/tsconfig-compass": "*",
+    "@types/async": "^3.2.7",
+    "@types/lodash": "^4.14.172",
     "bson": "^4.4.1",
     "chai": "^4.2.0",
     "depcheck": "^1.4.1",
