+++ packages/data-service/src/index.ts
@@ -1,0 +1,3 @@
+import DataService from './data-service';
+
+export = DataService;
