+++ packages/hadron-app/.depcheckrc
@@ -1,1 +1,9 @@
-ignores: []
+ignores: [
+    "babel-register",
+    "hadron-app-registry",
+    "less",
+    "mongodb-data-service",
+    "prop-types",
+    "react",
+    "react-dom"
+]
