--- packages/hadron-app/test/mocha.opts
@@ -1,1 +1,0 @@
---timeout 15000
